Introduced the substring matching search.
Now the "find" endpoint of the server can perform substring searches
both with the sqlAlchemy interface and the leveldb interface.
To achieve this, it will be sufficient to prepend a tilde to the query
string, e.g. searching for all hosts with a name containing "www" will be:

configdb-client find host --name="~www"

I still need to fix the docs, though, in order to reflect this api change.

Signed-off-by: Joe <joe@autistici.org>