etcd interface passing all tests, supports audit.
The etcd interface now passes all tests and also supports a very crude
audit logging. At the moment, only 100 audit messages are supported.

Due to a change of behaviour in dateutil, we had to change a little our
exception handling in the schema in order to pass tests in python
2.7.5.

The etcd driver has still not been road-tested